requirements.txt
gitpython
packaging